+++ packages/core/src/components/ozon-content/ozon-content.interfaces.ts
@@ -5,3 +5,7 @@ export interface OzonContentAnchorClick {
   documentComponent: string;
   originalEvent: MouseEvent;
 }
+
+export interface OzonContentClick {
+  originalEvent: Event;
+}